# services/API_DOCUMENTATION.yaml
openapi: 3.0.3
info:
  title: Apollo Platform API
  description: |
    ## Apollo Criminal Investigation Platform - Comprehensive API Documentation

    The Apollo Platform provides a complete suite of APIs for criminal investigation,
    intelligence gathering, and threat analysis. This documentation covers all available
    endpoints across Node.js microservices and Python intelligence services.

    ### Service Architecture
    - **Node.js Microservices** (Port 3000-3010): Authentication, Operations, Intelligence, Search, Analytics, Notifications
    - **Python Intelligence Server** (Port 8000): OSINT, Blockchain, Facial Recognition, Voice Analysis, GEOINT

    ### Authentication Methods
    - **JWT Bearer Token**: Primary authentication for all protected endpoints
    - **API Key**: Alternative authentication via `X-API-Key` header
    - **MFA**: Multi-factor authentication for sensitive operations

    ### Rate Limiting
    - Free tier: 100 requests/hour
    - Pro tier: 1,000 requests/hour
    - Enterprise: 10,000 requests/hour

    ### Error Handling
    All errors follow a consistent format with error codes and messages.

  version: 2.0.0
  contact:
    name: Apollo Platform Team
    email: dev@apollo-platform.com
    url: https://apollo-platform.com/support
  license:
    name: Proprietary
    url: https://apollo-platform.com/license
  x-logo:
    url: https://apollo-platform.com/logo.png
    altText: Apollo Platform
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid input provided"
            request_id:
              type: string
              example: "req_1234567890"
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "agent@apollo.com"
        username:
          type: string
          example: "agent_007"
        firstName:
          type: string
          example: "James"
        lastName:
          type: string
          example: "Bond"
        role:
          type: string
          enum: [admin, investigator, analyst, viewer]
          example: "investigator"
        clearanceLevel:
          type: string
          enum: [top_secret, secret, confidential, restricted, unclassified]
          example: "secret"
        isActive:
          type: boolean
          example: true
        isMfaEnabled:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "agent@apollo.com"
        password:
          type: string
          format: password
          example: "SecureP@ssw0rd!"
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            accessToken:
              type: string
              example: "eyJhbGciOiJIUzI1NiIs..."
            refreshToken:
              type: string
              example: "eyJhbGciOiJIUzI1NiIs..."
            expiresIn:
              type: integer
              example: 3600
...
    IntelligenceReport:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: "Suspicious Financial Activity Report"
        summary:
          type: string
        content:
          type: string
        source:
          type: string
          enum: [humint, sigint, osint, geoint, finint, techint]
        confidence:
          type: string
          enum: [verified, high, medium, low, unconfirmed]
        clearanceLevel:
          type: string
        authorId:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string
        relatedEntities:
          type: array
          items:
            type: object
        createdAt:
          type: string
          format: date-time

    # ============ Alert Schema ============
...
paths:
...
  /api/auth/login:
    post:
      summary: Login with email and password
      description: Authenticate user and receive JWT tokens
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "agent@apollo.com"
              password: "SecureP@ssw0rd!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid email or password"
        '403':
          description: MFA required
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  mfaRequired:
                    type: boolean
                    example: true
                  tempToken:
                    type: string
...
  /api/auth/refresh:
    post:
      summary: Refresh access token
      description: Get a new access token using refresh token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
                      expiresIn:
                        type: integer
...
  /api/intelligence:
    get:
      summary: List intelligence reports
      description: Get all intelligence reports
      tags: [Intelligence]
      security:
        - bearerAuth: []
      parameters:
        - name: source
          in: query
          schema:
            type: string
            enum: [humint, sigint, osint, geoint, finint, techint]
        - name: confidence
          in: query
          schema:
            type: string
            enum: [verified, high, medium, low, unconfirmed]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of intelligence reports
    post:
      summary: Create intelligence report
      description: Submit a new intelligence report
      tags: [Intelligence]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content, source, confidence, clearanceLevel, authorId]
              properties:
                title:
                  type: string
                summary:
                  type: string
                content:
                  type: string
                source:
                  type: string
                  enum: [humint, sigint, osint, geoint, finint, techint]
                confidence:
                  type: string
                  enum: [verified, high, medium, low, unconfirmed]
                clearanceLevel:
                  type: string
                authorId:
                  type: string
                  format: uuid
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Intelligence report created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/IntelligenceReport'
  /api/intelligence/{id}:
    get:
      summary: Get intelligence report
      description: Get intelligence report by ID
      tags: [Intelligence]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Intelligence report details
        '404':
          description: Report not found
...
  /api/v1/osint/email/intelligence:
    post:
      summary: Email intelligence
      description: Comprehensive email intelligence gathering
      tags: [OSINT]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Email intelligence results
...
security:
  - bearerAuth: []
  - apiKeyAuth: []